Extract telemetry package (Fixes #1589)
