+++ package.json
@@ -13,7 +13,8 @@
     "packages/a2a-server",
     "packages/test-utils",
     "packages/vscode-ide-companion",
-    "packages/lsp"
+    "packages/lsp",
+    "packages/telemetry"
   ],
   "private": "true",
   "repository": {
