+++ packages/core/package.json
@@ -30,6 +30,10 @@
     "./core/prompts.js": "./dist/src/core/prompts.js",
     "./debug/DebugLogger.js": "./dist/src/debug/DebugLogger.js",
     "./debug/index.js": "./dist/src/debug/index.js",
+    "./debug/ConfigurationManager.js": "./dist/src/debug/ConfigurationManager.js",
+    "./debug/FileOutput.js": "./dist/src/debug/FileOutput.js",
+    "./debug/MockConfigurationManager.js": "./dist/src/debug/MockConfigurationManager.js",
+    "./debug/MockFileOutput.js": "./dist/src/debug/MockFileOutput.js",
     "./models/hydration.js": "./dist/src/models/hydration.js",
     "./models/registry.js": "./dist/src/models/registry.js",
     "./models/provider-integration.js": "./dist/src/models/provider-integration.js",
@@ -52,6 +56,13 @@
     "./storage/secure-store.js": "./dist/src/storage/secure-store.js",
     "./telemetry/loggers.js": "./dist/src/telemetry/loggers.js",
     "./telemetry/types.js": "./dist/src/telemetry/types.js",
+    "./telemetry/index.js": "./dist/src/telemetry/index.js",
+    "./telemetry/sdk.js": "./dist/src/telemetry/sdk.js",
+    "./telemetry/constants.js": "./dist/src/telemetry/constants.js",
+    "./telemetry/metrics.js": "./dist/src/telemetry/metrics.js",
+    "./telemetry/uiTelemetry.js": "./dist/src/telemetry/uiTelemetry.js",
+    "./telemetry/tool-call-decision.js": "./dist/src/telemetry/tool-call-decision.js",
+    "./telemetry/file-exporters.js": "./dist/src/telemetry/file-exporters.js",
     "./test-utils/providerCallOptions.js": "./dist/src/test-utils/providerCallOptions.js",
     "./test-utils/mock-tool.js": "./dist/src/test-utils/mock-tool.js",
     "./test-utils/runtime.js": "./dist/src/test-utils/runtime.js",
@@ -91,6 +102,7 @@
   ],
   "dependencies": {
     "@vybestack/llxprt-code-mcp": "file:../mcp",
+    "@vybestack/llxprt-code-telemetry": "file:../telemetry",
     "@anthropic-ai/sdk": "^0.55.1",
     "@ast-grep/lang-c": "^0.0.5",
     "@ast-grep/lang-cpp": "^0.0.5",
@@ -111,15 +123,6 @@
     "@iarna/toml": "^2.2.5",
     "@lvce-editor/ripgrep": "^1.6.0",
     "@modelcontextprotocol/sdk": "^1.25.2",
-    "@opentelemetry/api": "^1.9.0",
-    "@opentelemetry/exporter-logs-otlp-grpc": "^0.203.0",
-    "@opentelemetry/exporter-logs-otlp-http": "^0.203.0",
-    "@opentelemetry/exporter-metrics-otlp-grpc": "^0.203.0",
-    "@opentelemetry/exporter-metrics-otlp-http": "^0.203.0",
-    "@opentelemetry/exporter-trace-otlp-grpc": "^0.203.0",
-    "@opentelemetry/exporter-trace-otlp-http": "^0.203.0",
-    "@opentelemetry/instrumentation-http": "^0.203.0",
-    "@opentelemetry/sdk-node": "^0.203.0",
     "@types/glob": "^8.1.0",
     "@types/html-to-text": "^9.0.4",
     "@xterm/headless": "5.5.0",
@@ -128,7 +131,6 @@
     "ajv-formats": "^3.0.0",
     "chardet": "^2.1.0",
     "cheerio": "^1.1.2",
-    "debug": "^4.3.4",
     "diff": "^8.0.3",
     "env-paths": "^4.0.0",
     "execa": "^9.6.0",
