+++ packages/core/src/debug/MockFileOutput.ts
@@ -1,37 +1,7 @@
 /**
- * @plan PLAN-20250120-DEBUGLOGGING.P05
- * @requirement REQ-001
- * Mock implementation for DebugLogger testing
+ * @license
+ * Copyright 2025 Vybestack LLC
+ * SPDX-License-Identifier: Apache-2.0
  */
-
-export interface LogEntry {
-  timestamp: string;
-  namespace: string;
-  level: string;
-  message: string;
-  args?: unknown[];
-  runId: string;
-  pid: number;
-}
-
-export class MockFileOutput {
-  private static instance: MockFileOutput | undefined;
-  private entries: LogEntry[] = [];
-
-  static getInstance(): MockFileOutput {
-    MockFileOutput.instance ??= new MockFileOutput();
-    return MockFileOutput.instance;
-  }
-
-  write(entry: LogEntry): void {
-    this.entries.push(entry);
-  }
-
-  getEntries(): LogEntry[] {
-    return [...this.entries];
-  }
-
-  clear(): void {
-    this.entries = [];
-  }
-}
+export { MockFileOutput } from '@vybestack/llxprt-code-telemetry/debug/MockFileOutput.js';
+export type { LogEntry } from '@vybestack/llxprt-code-telemetry/debug/MockFileOutput.js';
