+++ packages/core/src/utils/debugLogger.ts
@@ -1,15 +1,12 @@
 /**
  * @license
+ * Copyright 2025 Vybestack LLC
  * SPDX-License-Identifier: Apache-2.0
  *
- * Compatibility shim: upstream gemini-cli uses a singleton `debugLogger`
- * at `utils/debugLogger`. LLxprt uses the `debug`-based `DebugLogger`
- * class at `debug/DebugLogger`. This module exports a singleton instance
- * so that cherry-picked upstream code that imports `debugLogger` works
- * without modification.
+ * Compatibility shim: re-exports DebugLogger and the debugLogger singleton
+ * from the @vybestack/llxprt-code-telemetry package.
  */
-
-import { DebugLogger } from '../debug/DebugLogger.js';
-
-export { DebugLogger };
-export const debugLogger = new DebugLogger('llxprt:debug');
+export {
+  DebugLogger,
+  debugLogger,
+} from '@vybestack/llxprt-code-telemetry/utils/debugLogger.js';
