+++ packages/telemetry/package.json
@@ -1,0 +1,80 @@
+{
+  "name": "@vybestack/llxprt-code-telemetry",
+  "version": "0.10.0",
+  "description": "LLxprt Code Telemetry — debug logging, OpenTelemetry instrumentation, and metrics",
+  "license": "Apache-2.0",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/vybestack/llxprt-code.git"
+  },
+  "type": "module",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    },
+    "./debug/index.js": "./dist/src/debug/index.js",
+    "./debug/DebugLogger.js": "./dist/src/debug/DebugLogger.js",
+    "./debug/ConfigurationManager.js": "./dist/src/debug/ConfigurationManager.js",
+    "./debug/FileOutput.js": "./dist/src/debug/FileOutput.js",
+    "./debug/MockConfigurationManager.js": "./dist/src/debug/MockConfigurationManager.js",
+    "./debug/MockFileOutput.js": "./dist/src/debug/MockFileOutput.js",
+    "./utils/index.js": "./dist/src/utils/index.js",
+    "./utils/debugLogger.js": "./dist/src/utils/debugLogger.js",
+    "./utils/safeJsonStringify.js": "./dist/src/utils/safeJsonStringify.js",
+    "./utils/session.js": "./dist/src/utils/session.js",
+    "./telemetry/loggers.js": "./dist/src/telemetry/loggers.js",
+    "./telemetry/types.js": "./dist/src/telemetry/types.js",
+    "./telemetry/index.js": "./dist/src/telemetry/index.js",
+    "./telemetry/sdk.js": "./dist/src/telemetry/sdk.js",
+    "./telemetry/constants.js": "./dist/src/telemetry/constants.js",
+    "./telemetry/metrics.js": "./dist/src/telemetry/metrics.js",
+    "./telemetry/uiTelemetry.js": "./dist/src/telemetry/uiTelemetry.js",
+    "./telemetry/tool-call-decision.js": "./dist/src/telemetry/tool-call-decision.js",
+    "./telemetry/file-exporters.js": "./dist/src/telemetry/file-exporters.js"
+  },
+  "scripts": {
+    "build": "node ../../scripts/build_package.js",
+    "lint": "eslint . --ext .ts,.tsx",
+    "format": "prettier --write .",
+    "test": "vitest run",
+    "test:ci": "vitest run",
+    "typecheck": "tsc --noEmit"
+  },
+  "files": [
+    "dist"
+  ],
+  "dependencies": {
+    "@google/genai": "1.30.0",
+    "@opentelemetry/api": "^1.9.0",
+    "@opentelemetry/api-logs": "^0.203.0",
+    "@opentelemetry/core": "^1.26.0",
+    "@opentelemetry/exporter-logs-otlp-grpc": "^0.203.0",
+    "@opentelemetry/exporter-logs-otlp-http": "^0.203.0",
+    "@opentelemetry/exporter-metrics-otlp-grpc": "^0.203.0",
+    "@opentelemetry/exporter-metrics-otlp-http": "^0.203.0",
+    "@opentelemetry/exporter-trace-otlp-grpc": "^0.203.0",
+    "@opentelemetry/exporter-trace-otlp-http": "^0.203.0",
+    "@opentelemetry/instrumentation-http": "^0.203.0",
+    "@opentelemetry/resources": "^2.0.0",
+    "@opentelemetry/sdk-logs": "^0.203.0",
+    "@opentelemetry/sdk-metrics": "^2.0.0",
+    "@opentelemetry/sdk-node": "^0.203.0",
+    "@opentelemetry/sdk-trace-base": "^2.0.0",
+    "@opentelemetry/sdk-trace-node": "^2.0.0",
+    "@opentelemetry/semantic-conventions": "^1.26.0",
+    "debug": "^4.3.4"
+  },
+  "devDependencies": {
+    "@types/debug": "^4.1.12",
+    "@types/node": "^24.2.1",
+    "fast-check": "^4.2.0",
+    "typescript": "^5.3.3",
+    "vitest": "^3.1.1"
+  },
+  "engines": {
+    "node": ">=20"
+  }
+}
