+++ packages/telemetry/src/debug/index.ts
@@ -1,0 +1,10 @@
+/**
+ * @plan PLAN-20250120-DEBUGLOGGING.P03
+ * @requirement REQ-001
+ */
+export * from './types.js';
+export * from './DebugLogger.js';
+export * from './ConfigurationManager.js';
+export * from './FileOutput.js';
+export { MockConfigurationManager } from './MockConfigurationManager.js';
+export { MockFileOutput } from './MockFileOutput.js';
