+++ packages/telemetry/src/utils/paths.ts
@@ -1,0 +1,11 @@
+/**
+ * @license
+ * Copyright 2025 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * The directory name used for LLxprt configuration and data.
+ * Duplicated here to avoid importing from core.
+ */
+export const LLXPRT_DIR = '.llxprt';
